// apps/web-svelte/src/lib/api/generated/model/customFormatView.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { CustomFormatConditionView } from './customFormatConditionView';
import type { EntityKind } from './entityKind';

export interface CustomFormatView {
  id: string;
  kind: EntityKind;
  name: string;
  conditions: CustomFormatConditionView[];
}